test: cleanup refactor
