+++ packages/tests-utils/src/package-manager.ts
@@ -1,0 +1,4 @@
+import which from "which";
+
+export const bunExists = which.sync("bun", { nothrow: true }) !== null;
+export const npmCli = bunExists ? "bun" : "pnpm";
