+++ packages/tests-utils/src/port.ts
@@ -1,0 +1,6 @@
+import getPort from "get-port";
+import type { GlobalContext } from "./types.js";
+
+export async function initPort(context: GlobalContext) {
+  context.port = await getPort();
+}
