+++ packages/tests-utils/tsconfig.json
@@ -6,5 +6,5 @@
     "preserveSymlinks": true,
     "types": ["@types/node"]
   },
-  "include": ["src/*", "tests/*"]
+  "include": ["src/*"]
 }
