+++ packages/tests/src/types.ts
@@ -1,0 +1,3 @@
+export interface GlobalContext {
+  tmpdir: string;
+}
