+++ packages/tests/tests/FRAMEWORK+SERVER+telefunc.spec.ts
@@ -1,4 +1,4 @@
-import { describeBati } from "./utils";
+import { describeBati } from "@batijs/tests-utils";
 
 export const matrix = [["solid", "react", "vue"], ["express", "hattip", "h3"], "telefunc"];
